server/requirements.txt
fastapi
uvicorn
opencv-python
numpy
paho-mqtt
Pillow